Use list of words that `common/nbsp/afterShortWord` should only be applied to
The `common/nbsp/afterShortWord` rule only considers the word length
to determine if it is short or not, without taking into accounts its semantics
(whether it is a content or function word).
This commit adds a `useShortWordList` setting whereby the rule can use a predefined list of short words.
Two lists of function words are provided for the `ru` and `en-US` locales.